tests/wpt/web-platform-tests/cssom-view/MediaQueryList-001.html

<!DOCTYPE html>
<html>
<head>
    <title>CSSOM View Module test:MediaQueryList</title>
    <link rel="author" title="unbug" href="mailto:tidelgl@gmail.com" />
    <link rel="help" href="http://www.w3.org/TR/cssom-view/#the-mediaquerylist-interface">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style type="text/css">
    </style>
</head>
<body>
    <p>This case tests the MediaQueryList
    <p>The test passes if the value is 1280/800</p>
    <div id="log"></div>
    <script>
        test(function(){
            assert_equals(window.matchMedia('(device-aspect-ratio: 1280/800)').matches, true, "Expected value for device-aspect-ratio is 1280/800");
        },'matchMedia');
    </script>
</body>
</html>
