third_party/WebKit/LayoutTests/external/wpt/cssom-view/MediaQueryList-001-expected.txt

This is a testharness.js-based test.
FAIL matchMedia assert_equals: Expected value for device-aspect-ratio is 1280/800 expected true but got false
Harness: the test ran to completion.
